Update Invoice + Fee/Fund
