+++ src/features/invoices/InvoiceForm.tsx
@@ -1,0 +1,9 @@
+
+
+const InvoiceForm = () => {
+  return (
+    <div>InvoiceForm</div>
+  )
+}
+
+export default InvoiceForm
